+++ index.html
@@ -27,6 +27,10 @@
           id="municipio-select"
           aria-label="Selecionar município"
         ></select>
+        <!-- Spinner para feedback visual enquanto os municípios estão sendo carregados -->
+        <div id="municipio-spinner" style="display: none">
+          <span class="spinner"></span> Carregando...
+        </div>
         <button id="consultar-btn">
           Consultar
           <span
@@ -53,9 +57,9 @@
       </div>
     </section>
 
-    <main>
+    <main role="main">
       <h1 id="title">Previsão do Tempo</h1>
-      <div id="cards-container"></div>
+      <div id="cards-container" aria-live="polite"></div>
     </main>
     <!-- Botão voltar ao topo -->
     <div id="back-to-top" class="floating-btn" aria-label="Voltar ao topo">
@@ -73,6 +77,6 @@
         <polyline points="6 15 12 9 18 15"></polyline>
       </svg>
     </div>
-    <script src="js/app.js"></script>
+    <script src="js/index.js"></script>
   </body>
 </html>
