--- js/storage.js
@@ -1,18 +1,0 @@
-export const Storage = {
-  getHistory: () => JSON.parse(localStorage.getItem("weatherHistory")) || [],
-  saveHistory(city, maxItems = 5) {
-    const formattedCity = city.charAt(0).toUpperCase() + city.slice(1);
-    let history = this.getHistory().filter(c => c.toLowerCase() !== formattedCity.toLowerCase());
-    history.unshift(formattedCity);
-    localStorage.setItem("weatherHistory", JSON.stringify(history.slice(0, maxItems)));
-  },
-
-  getFavorites: () => JSON.parse(localStorage.getItem("weatherFavorites")) || [],
-  saveFavorites(favorites) { localStorage.setItem("weatherFavorites", JSON.stringify(favorites)); },
-
-  getTheme: () => localStorage.getItem("theme") || "light",
-  saveTheme: theme => localStorage.setItem("theme", theme),
-
-  getLastCity: () => localStorage.getItem("lastCity"),
-  saveLastCity: city => localStorage.setItem("lastCity", city),
-};
